[kitchen] Make -workdir shorter.
This is to allow as many characters as possible to Windows tasks which
can run into MAX_PATH issues.

This also makes the recipe subcommand cd into the workdir as well;
currently the recipe engine picks a crazy CWD if `--workdir` is
unspecified:

  https://chromium.googlesource.com/infra/luci/recipes-py/+/master/recipe_engine/run.py#459

By changing `Dir`, it opens up the opportunity for us to remove this
flag from the recipe engine once buildbot is deprecated.

R=nodir@chromium.org, tandrii@chromium.org

Change-Id: I95dc64c59f66d9e4578990b540cb2ed2d6b0769f
Reviewed-on: https://chromium-review.googlesource.com/c/1446288
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20450}