[vpythonsmoketest] Add a simple smoke test
The idea is to figure out any race condition in vpython that could cause
the linked bug.

It is very simple for now, will expand as more test cases are thought of.

The smoketest is fast enough that it's run as part of 'go test'.

Bug: 869227
Change-Id: I3d436609a6420db1f281aa63362a89cc8a733d3b
Reviewed-on: https://chromium-review.googlesource.com/c/1459679
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20586}