qscheduler: serialize labels as integers, not strings
The bulk of the serialized storage size of a scheduler object is the
many label strings, most of which are repeated frequently among tasks
and workers many times.

Shrink the serialized size by storing labels as integers with a
state-local lookup map, rather than as raw strings.

BUG=925546

Change-Id: Id93b81d249d2de0b7c638ed7f49c6ee86c66231f
Reviewed-on: https://chromium-review.googlesource.com/c/1437938
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20662}