[tricium] In spacey and spellchecker, handle long lines
Before this change, upon encountering a file with a long
line (e.g. a minified JS file), both of these analyzers
would just fail and thus have no results.

Instead, this would make it so we just skip the problem
file and report results for the rest of the files.

Bug: 926589
Change-Id: I98cac86c304f7a44e4c1d29b8792927e8de2f29a
Reviewed-on: https://chromium-review.googlesource.com/c/1473930
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20686}