[Monorail] Block user endlessly polling v8-related issues.
User seems to poll a set of issues every 15 or 30 minutes.  That itself
is probably not too harmful to other users, but the set of issues being
polled are particularly long ones with lots of comments containing
lots of links, which are slow to serve.

R=ehmaldonado@chromium.org

Change-Id: Ib2143dc6da7c0d86ad2c182caea7577a0acc98d3
Reviewed-on: https://chromium-review.googlesource.com/c/1473932
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Auto-Submit: Jason Robbins <jrobbins@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20691}