[Findit] Fix flake detection queries to filter out tests with 'NOTRUN' or 'UNKNOWN' results.
Test runs with result 'NOTRUN' or 'UNKNOWN' are nondeterministic so we cannot use. And Tests with those results tend to happen in a large group, indicating there's something wrong with test runner or infra, rather than individual tests. So we decide to ignore the tests entirely to reduce noice.

Bug: 929442
Change-Id: I72f2226a3495439c1e9fece00ab64fa58798bb40
Reviewed-on: https://chromium-review.googlesource.com/c/1474339
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20706}