[Findit] Handle 'NOTRUN'/'UNKNOWN' test result statuses.
'NOTRUN' is a placeholder in gtest results before any tests runs, and when each test runs, the status will be updated to real results. So if for any reason a test doesn't run eventually, the test result for it will remains to be 'NOTRUN'.
'UNKNOWN' status on the other hand is a result for a test has run yet the test result is unknown.

1. At Findit side, add a new test result classification 'notruns'.
  a. For gtests, classify test result in 'notruns' if the result status is 'NOTRUN'
  b. 'NOTRUN' is not meaningful for webkit_layout_test, the 'notruns' classification will always be empty
2. Ignores 'notruns' when processing swarming rerun results, same as unknown tests.

Bug: 929442
Change-Id: I16b621a564a5cabbefcdd50f10681c911dde2eb2
Reviewed-on: https://chromium-review.googlesource.com/c/1459684
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20709}