[dockerbuild] Fix bug where tags were assumed to be versions
Dockerbuild assembles a couple different tags to attach to packages, one of them
being the git revision of infra from which the wheel was uploaded.

Unfortunately, when instructed to upload the wheel, it considers this
git_revision tag to be an authoritative version. This means that if you add
multiple versions of a package from the same commit then only one of them will
actually be uploaded and tagged correctly.

This cl will now explicitly look for the "version:" tag and use that as the
authority for whether or not it should attempt to upload the wheel.

R=nodir@chromium.org

Change-Id: I965507e2ae1795996759a996eaf05ec15726fdab
Reviewed-on: https://chromium-review.googlesource.com/c/1475239
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20727}