[Monorail] Make card borders faint as a step toward eliminating.
I think we overstepped on the card motif, but I don't want to reverse
it so soon because we don't want to give users change-fatigue.  So,
this CL makes the cards less noticable, and we'll follow up in a few
weeks with another CL that fully eliminates the borders.

R=zhangtiff@chromium.org

Bug: monorail:4772
Change-Id: I19f4b0dc6c2c7ff3d7958dc21a7d50fcd9679034
Reviewed-on: https://chromium-review.googlesource.com/c/1474543
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20731}