[tricium] Add logging when fetching buildbucket builds
Purpose: debugging

Background: Recently Julie enabled recipe-based tricium
runs for tricium-prod, which appear to be working up until
the point when tricium fetches the output. The output is in
the "tricium" field of the output properties.

This output had been previously successfully fetched, but
now it appears that

  `build.GetOutput().GetProperties().GetFields()["tricium"].GetStringValue()`

is empty string, even though on milo there does appear to be a
"tricium" field in the output properties.

Bug: 931768
Change-Id: Ida185dca732781e1f0ab777e11ecf24d55f4247f
Reviewed-on: https://chromium-review.googlesource.com/c/1478013
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20776}