qscheduler: use float32 instead of float64 for balance tracking
float32 provides 23 bits of precision (~1 part in 8 million) which is
sufficient for tracking quota and costs (measured in DUT seconds,
typically ranging from 0-100,000, and updated in increments typically
exceeding 1 DUT second).

This is a proto-breaking change, but since the rollout of quotascheduler
is still small it can be accomplished with some babysitting and no user
visible issues.

BUG=925546

Change-Id: I7ee1a2640473564de99ac023c0c267b531cf1743
Reviewed-on: https://chromium-review.googlesource.com/c/1479260
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20801}