Re-add the Paths field to track.AnalyzeRequest
Paths is now unused but exists on older entities;
it must be retained in the struct so that older entities
can still be loaded.

I chose to re-add it here rather than handling the
particular case of failing to load an AnalyzeRequest in
the Progress RPC because there may be other cases where
we want to load older entities, e.g. maybe sending backlog
of analyze results to BQ.

Bug: 934246
Change-Id: I6f0e4c0696b3ae2bf58e9774a6177b0c0f1fb0ad
Reviewed-on: https://chromium-review.googlesource.com/c/1481456
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20823}