[Findit] Fix the bug to make sure after a flake bug is created for a group, a follow-up comment will be posted with links.
Findit is supposed to send a follow-up comment after a flake bug for group is created. Because of a bug in code the comment was not sent.

And an unnessary mock in test mask the bug.

This change should fix the code and the test.

Bug: 922235

Change-Id: I406926ff041efacd222fb6fa692f772d5fa95125
Reviewed-on: https://chromium-review.googlesource.com/c/1483781
Reviewed-by: Shuotao Gao <stgao@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20870}