[tricium] Separate merging and validating of configs
A recent issue involving storing project configs with no repos set was
caused by using the returned project config from Validate, which was
not a complete config but in fact a partial config with merged functions
that was used in workflow generation.

This CL would change Validate to just return an error or nil,
and would change the workflow generation to just get the merged
functions without necessarily validating other parts of the
configs (which should now be validated at ingestion time).

Also in this CL:
 - Rename variable names to reflect the (long-ago)
   rename of "analyzer" to "function", e.g. sa -> sf
 - Rename mergeFunctions -> mergeFunction
 - Remove some now-unnecessary logging
 - Edit comments

Bug: 932213
Change-Id: I1987f6be55659bde6cea44bbf09536d820ba7260
Reviewed-on: https://chromium-review.googlesource.com/c/1484712
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20876}