crosskylabadmin: [Breaking] change in deployment API.
This API is not yet implemented, so is clearly unused.
There isn't going to be more than one task. And if there is in the
future, its more meaningful to include the tasks separately so that
users know which task to follow for which step.

BUG=chromium:929776

Change-Id: I9b95de9c45b58aa7bf631ec74bc57c4666e832eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1491941
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#20995}