qscheduler: handle NotifyTasks errors internally to operation
Any per-item errors that occur within a NotifyTasks handler should not
be returned to the caller. Per-task-request errors are instead stored so
that a subsequent GetCancellations call with cancel erroneous tasks.

Handling errors in this way also ensures that the NotifyTasks operation
is "atomic". It either returns an error right away without mutating
state, or it applies all its mutations without returning an error.

BUG=923515

Change-Id: I6968d751f4db981da88168735cb8ebe7be3ef4f7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1496381
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21036}