[tricium cpplint] Add extra processing for include_what_you_use warnings
This is one proposed way of handling a category of cpplint comments
that appears to have a relatively high false positive rate (just based
on bug reports). It's not very feasible fix cpplint, so we could filter
and amend the messages from cpplint. There is precedent for this as we
do this for pylint.

An alternative possible option would be to just disable this check
for chromium/src.

Bug: 935671, 936259
Change-Id: I87f8176e8cd20070d92f313f7c6cb61d7126b3ab
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1500642
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21114}