[tricium] Correctly set the CommentSelection.included field.
Previously rpc_worker_done defaulted all comment selections
to true and would post a task to gerrit/internal/report-results
which was supposed to filter out all comments not in the Gerrit
patchset. Unfortunately the gerrit/internal/report-results
handler never updated the datastore.

This change sets the CommentSelection.included field before the
first write to the datastore and Gerrit/internal/report-results's
only job is to report those comments to Gerrit.

Making the request to Gerrit in rpc_worker_done was necessary
because the value needs to be correctly set before writing the
results to BigQuery.

Bug: 937385
Change-Id: I8138cee26ecc38d91b323f48f4bdf1ed85c2d2ba
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1499788
Commit-Queue: Chris Mumford <cmumford@google.com>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21134}