[tricium eslint] Change setup.py fetching of node and node_modules
After re-imaging my workstation, setup.py didn't work
as it had before; eslint would be installed in a different
location rather than locally.

This change adds a package.json listing the required
dependencies instead of listing them in a command
invocation in the script.

I believe this change should fix setup.py so that the
required node_modules are put in this same directory.

This also changes the script to not download node if
it's already been fetched.

Bug: 938002
Change-Id: I45049846434b955974029038c32eb82274f09b84
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1509114
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21185}