[cr-buildbucket] Start recognizing X-Luci-Project headers.
Still fallback (with logging) to checking e.g. 'luci-scheduler@...' is in the
ACL, so this CL doesn't break everything.

Another caveat is that PermittedActions RPC does not recognize X-Luci-Project
and everything will break horribly if callers (Milo?) will try to pass
X-Luci-Project. But this should be fine, since PermittedActions is supposed to
be used ONLY when checking end-user's roles and it relies on delegation tokens
instead of X-Luci-Project (these two mechanism are incompatible and impossible
to use together).

R=nodir@chromium.org
BUG=938083

Change-Id: Ie77c0056d3906c519be4c4545092ae5730f836ff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1519648
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21412}