Monorail: Fix ts mon metrics logging for issue detail and issue list views.
Demo: https://21535-52ba193-tainted-zhangtiff-ts-mon-dot-monorail-staging.appspot.com

The gist of this problem was:

When Sean and I added ClientLogger instances to the header, these ClientLogger instances initialized the MonorailTSMonClient instance.
Before the header used ClientLogger, ClientLogger was only initialized within EZT inside of runOnLoad.
Since MonorailTSMonClient ran performance recordings in its constructor, this means that when ClientLogger inside the header created a new MonorailTSMonClient instance, now MonorailTSMonClient was recording before the page was ready.
To make it less likely for this sort of breakage to happen in the future, I changed MonorailTSMonClient slightly so that it no longer starts recording inside of its constructor and instead moved to a model where pages explicitly handle their own performance recording. I believe this will be more flexible when we move more metrics to the SPA and will be less likely to break unexpectedly in the future.

Bug: monorail:5118
Change-Id: Ibac9e098809eea9d0a13f883fd87d40defe1a204
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1525553
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21501}