[Monorail] Fix "Back to list" URL after jumping to issue ID.
This avoids a 500 in the case where a user searches for an issue ID,
e.g., 1234 that causes them to jump directly to the issue, and then
clicks the "Back to list" link.  This CL causes that link to search
for [id=1234] which will be a list of one item.

R=ehmaldonado@chromium.org

Bug: monorail:5164
Change-Id: Ib1fe35c374fde2d50e5f4a301d650debb3a3d948
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1529049
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21578}