skylab_swarming_worker: Add ADMIN_SERVICE bot info
This should be the proper way to pass the admin service setting since
it is part of the environment the worker is running in.  Previously
this was only going to be used for label updates, but now that we will
use this for getting dut info for every task, it makes more sense as
an env.

Keep supporting the -admin-service flag while transitioning.

R=akeshet@chromium.org

Bug: 941159
Change-Id: I8bc018638d235516513be79abba39cf576e73722
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1530013
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21619}