Initial commit for Arquebus with config and frontend packages.
Arquebus is an automatic issue assignment service (go/arquebus) for Monorail,
and the project has been developed and tested with/in
- https://arquebus-staging.appspot.com
- the code in http://crrev.com/i/877748

I've broken http://crrev.com/i/877748 into a series of CLs to make reviews easier,
and there wil be a couple of follow-up CLs after this.

This CL includes the following packages

@1. appengine
An appengine entry point, where init() imports and invokes the hooks of all
other packages.

@2. frontend
An HTTP frontend rendering UI and processing JSON requests. In this CL,
it simply sets up auth channels and ACL checks.

@3. config
Defines the proto of Arquebus configuration data. It also installs a hook
to manage the configuration object that is accessible by all other packages.


R=tandrii

Bug: 849469
Change-Id: I59b8c4a7dead03ee03ba6892ff484a3d0bec039c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1530415
Commit-Queue: Scott Lee <ddoman@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21653}