[rotang] Add in a verification for TZFair.
TZFair schedules a member to be on-call per TZ.
This CL adds a verification to check that number of TZs in a rotation are
the same as number of Sheriffs to schedule.

Snuck in some more logging for the blank EvtID in here too.

Bug: 943008
Change-Id: I214afe2b65f43d1316f66bdb17a1d53bae3dfcad
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1530412
Reviewed-by: Scott Lewis-Kelly <slewiskelly@chromium.org>
Commit-Queue: Ola Karlsson <olakar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21666}