Add basic dataflow to join against buildbucket.
This CL adds a new dataflow that extends the functionality of cq_attempts. In
addition to sanitizing the CQ events, this workflow joins the results against
buildbucket to set the failure type to PRESUBMIT_FAILURE if the only failures
are caused by 'chromium_presubmit'.

Bug: 937519
Change-Id: I4d4b8c61667b2633a1f65037ee3acdd5a66e7daa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1521042
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21771}