[Monorail] Serve JS and CSS from VERSION-dot-APPID.appspot.com.
Now, a request to get JS will always go to the version of the app that
corresponds to the version that generated the HTML page.  We no longer
need the ?version= parameter in the URL because the host varies with
each deployment.  (To easily see the host in the GAE logs, add a column
for protoPayload.host.)

The CORS header is needed for JS code loaded as a module.

R=zhangtiff@chromium.org

Bug: monorail:5248
Change-Id: I2d155f16bffe2edf2b7ecb40b884d54427e37f7e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1542197
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21847}