Dont fail cros dut flash scheduler runs if a flash fails.
The logic being removed here is redundant since we already test the
exit code of the flashing job. We don't also need to check if the OS
gets updated too.

In other words, it's safe to say that the flashing task will exit nonzero
if the DUT doesn't get updated.

Bug: 866062
Change-Id: I2b4589c48dbd2d8313b34f4a53ac0f976f89244d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1542096
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21862}