[Monorail] Remove saved queries from servlet.py
These values used to be needed on virtually all pages to populate the
EZT version of the search scope widget.  However, now that is done
via an API call, so it is not needed in the initial generation of
most pages.  It is only needed on the pages that edit project and
user saved queries.

This saves about 3ms for most users and probably 10ms for some users.
That is not a big latency savings, but it is very consistent across
all page requests.  And, removing this cleans up some code and avoids
so logs clutter.

R=ehmaldonado@chromium.org

Bug: monorail:5018
Change-Id: I8c431d1773b6e19a7071367b102b5f50f5275049
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1544934
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21922}