crosskylabadmin: Factor out DUT drone lookup
This has one behavioral change; trying to remove a DUT from
the wrong drone is changed from no error to an error.  The rational is
that if the request provided a drone, the intent is to check to make
sure everything is as expected.  If things are not as expected, fail
loudly.

R=akeshet@chromium.org

Bug: 934067
Change-Id: I5aa0a50349b99e7ac2b7bb2012e668de95ca0fe2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1540572
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21935}