[monorail] Combine issueId and projectName into issueRef.
Also rename issueId to localId whenever it's encapsulated in issueRef.
This is consistent with the name in the API and simplifies prpc calls.

Exceptions: ezt-comments.js, mr-keystrokes.js
The parents of these components collect projectName and issueId
from separate places, so it's not convenient to combine them.

Bug: monorail:5135
Change-Id: Ib1e1609278abdd5c651f3c4434c0e22c708eeaf9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1544937
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21940}