crosskylabadmin: Clean up removeDutFromServer
There are a lot of functions here called something like removeDut.  In
context, it's hard to tell what this function does exactly.

Fix that by partially inlining it and adding an obviously named
removeSliceString helper, so that it is obviously just mutating a
slice in place.

R=akeshet@chromium.org

Bug: 934067
Change-Id: I7f7a62ec12cc0df0d2929485a5c88bd4a9ddf519
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1544936
Commit-Queue: Allen Li <ayatane@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21950}