Bump infra_libs version to 2.0.5
This is to release a new infra_libs package with the patch of
crrev.com/c/1545615. The intent of this release is the same as of 2.0.3.
However, 2.0.3 was abandoned before it gets rolled out because of 2.0.4.
crrev.com/c/1551826.

2.0.4 has been rolled out, and this CL is to release 2.0.5 with the patch
of crrev.com/c/1545615 again.


R: tandrii
Bug: 862894
Change-Id: I4793138feda2774285e76d6cfcf356bb1e93df36
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1552199
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Commit-Queue: Scott Lee <ddoman@google.com>
Cr-Commit-Position: refs/heads/master@{#22103}