[Monorail] End backward compatability with pRPC trace field.
We used to pass the XSRF token in request.trace.token as part of each
request proto.  Now, it is passed as a HTTP header.  The JS in our app
that makes those requests has been using the HTTP header for two weeks.
So, this CL removes support for using the trace proto and also removes
the whole trace proto.

R=ehmaldonado@chromium.org

Bug: monorail:5175
Change-Id: Ibc1d0bebdc74ff284feb12e6816d28d02a2bd64e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1559294
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22256}