skylab: Accept DeviceUnderTest specs for inventory updates
CommonDeviceSpecs are common to all device types. The `add-dut` and
`update-dut` subcommands are intended to be used only for DUTs, so use
DeviceUnderTest spec at the CLI.

This also makes it possible to create a pipeline that uses `skylab
dut-info -json` to obtain specs that are then provided to the inventory
update subcommands.

Bug: 912977
Change-Id: I42fc3fc5d204682e03aa2b91479b8c73e14457e1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1564658
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22353}