[monorail] FLT: Show issue page before comments are loaded.
Showing the issue page is on the critical path - we don't load
anything else until it's attached to the DOM. Fetching the comments
can take a while, so it shouldn't be in the critical path.

This change allows us to start some work earlier, which reduces
the total page load time by around 100ms for 200 comments.

Bug: monorail:5319
Change-Id: I8ce7654f51eec574b3afb5955a9c5a461eb2b1bc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1553298
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22355}