[monorail] FLT: defer comment rendering for hidden comments.
When there are more than 100 comments on an issue, the oldest ones are
hidden behind a link. Currently, we still do the work to attach them
to the DOM at page load time.

I *think* this saves 600 ms for 100 hidden comments.
But this will also be less with other perf improvements we're making.

Bug: monorail:5319
Change-Id: Iaf704972a1847f62ea07744e978159e80fa1ea9d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1561896
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22486}