qscheduler: add an Unassign method to scheduler
It is possible for the quotascheduler to assign a task to a worker, but
for that worker to never pick it up. In such instances, we need a way to
move the task back to the queue (after a timeout) so that it is not
lost.

This CL adds an Unassign method to the scheduler which can be used to do
so.

BUG=953783

Change-Id: I165c6aa0f725ae7d7a64a2861182fbf02780a261
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1575015
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22551}