[som] Update command used to generate relnotes
The old command didn't use the "-extra-paths" flag, and it passed
the paths to check as a space-separated list, causing them to be
read as positional args instead of flag args. This makes it impossible
to add additional flags (such as -since-hash), and causes the command
to print errors as it tries to ask git to parse paths it doesn't
recognize.

Change-Id: I0d4a98f8b4e85700eae76c85dc055c4e61954046
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1575008
Auto-Submit: Aaron Gable <agable@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22640}