[rotang] Add in Schedule generators `Recent` and TZRecent.
The fair generator does not work well for the Build Sheriff rotation.
The Recent generators only considers how recent a member was on-call.

This gives if a member was just on-call they'll not be scheduled again
before all others have had a shift.
If there are multiple members who don't have any previoius shifts in history
a random member will be choosen.

Bug: 958465
Change-Id: I253698cafcc55fd16f965d797af146e582bd9113
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1592574
Commit-Queue: Ola Karlsson <olakar@chromium.org>
Reviewed-by: Scott Lewis-Kelly <slewiskelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22696}