skylab: create-test: add -parent-task-id argument
To make use of swarming's parent-child relationships, child tasks need
to be created with a parent task id. This is the first step in allowing
this change, and will be followed up by a change to run_suite_skylab
that makes use of this argument.

BUG=958892
TEST=None

Change-Id: Iadb3ec76d7ecb3242b44e206b2d43c2205f17fb9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1593686
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22706}