[mr] Run frontend tests in parallel.
This uses the karma-parallel plugin, which in turn requires us to use
bdd instead of tdd (which amounted to string substitution,
so tdd ~= bdd with different method names AFAICT). Hence every test
got touched.

This hard-codes the number of shards to 7, since we run CI on an 8 core
VM.  This way you'll see the same sharding locally as you do on CQ.

One downside is that you'll now get 7 browser windows open when you run
`make karma` locally, but it's easy enough to cut the shards down to
1 if you need to debug something.

Bug: monorail:5264
Change-Id: I810151da83f63e5d5c1dc6fd9817f5f127f86001
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1600511
Commit-Queue: Sean McCullough <seanmccullough@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#22827}