[Findit v2] Add logic to group compile failures in ChromeOS builds.
When Findit gets a chromeos build with some compile failures that start to fail from that build, it'll look for existing failure groups in which the failures are highly likely caused by the same culprit.

The criteria for a matching group are:
+ Same LUCI project
+ same regression range (last_passed_commit to first_failed_commit)
+ exactly the same failed targets (partial match doesn't count)

If such failure group is found, Findit will merge the failures in the new build to them, otherwise start a new analysis.

Bug: 949836
Change-Id: I52febec51fa3db511cd04efd8aec6af4e5c5c7d5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1607159
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23004}