[rotang] Schedule extension job feeding only current shifts to the generator.
This CL fixes a bug in the Schedule Extend job causing some people in the Build
Sheriff rotation to be scheduled too often.

Cause is the Schedule Extension job using only the current shifts when calling the
Generator.
With this change the full history is fed to the generator.


Bug: 953539
Change-Id: I5d22281a0c4e71352bd64e165fbfa014a89cab06
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1627172
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Ola Karlsson <olakar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23176}