skylab: add task-run-id to wait-task result
In some cases it may be necessary to know which try # of a request was
run (i.e. which swarming run id the result corresponds to). Most
commonly, in the case of fetching child results (from tko) from a suite
job. Add a task-run-id to wait-task to make this possible to determine.

BUG=966518
TEST=local test

Change-Id: I7068451ff925bc36649076135e98ae2224f55eeb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1628127
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23184}