[Findit v2] Add merged_failure_key to CompileFailure model
merged_failure_key is the key to the failure the current failure merges into.

For example, if the same failure happens on build 110 and 111, then the failure on build 111 will merge into the build on 110;
Then if the same failure also happens on build 500 on another builder and Findit already gets it, failures on both 110 and 111 will merge into it.

With this info, it'll be easier to retrieve analysis results for failures - no matter if the failure is actually analyzed or it's using another analysis's result.

Bug: 943744
Change-Id: I163e2839f1d7e31fa1f0a82c9e4d1fddb8a9300f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1650743
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23573}