[monorail] Add issue detail SPA ts_mon metrics.
We want to measure two metrics:
* Time from Navigation Start to Comment List, for when the user
  is navigating into the SPA from outside. ("from outside app")
* Time from Click to Comment List, for when the user
  is navigating within the SPA. ("within app")

These metrics do not include the non-JS work: style, layout, paint, composite,
which may be ~10-100 ms based on the number of comments.

For marking the "Click", it's hard to keep track of all the links that navigate
to other issues, so instead I've used the start of the issue detail page load.
This is ~1-10 ms after the click.

Marking when the comment list is rendered is surprisingly difficult. I've used
MrIssueDetails.updated() because there is only one on the issue detail page.
The number and order of updates vary based on the order of network responses, so
I've added some filters to pinpoint only the cases we care about. These filters
may or may not prove to be brittle in the future. lit-element updates as also
asynchronous, so I recursively collect all the child nodes and wait for all of
them to finish updating.

Bug: monorail:5565
Change-Id: Id6a77833974563f9cd6b65034a92af590a64746e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1650135
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Jeff Carpenter <jeffcarp@chromium.org>
Commit-Queue: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23694}