[buildbucket] Cleanups
Update docstrings stating that functions mutate build.
This no longer happens, and if it does, these changes are not saved
anywhere.

Stop writing service_config_revision, since it is not saved.
We want to restore it later, but it is not used for anything too

Remove useless cloning.

R=tandrii@chromium.org

Change-Id: I291a315fe88cd4e5a0ef671a7521ff8b36361775
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1661046
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23735}