[tricium] Add --uses-shflags to shellcheck
We added a --uses-shflags argument to shellcheck in 0.6.0-r6 that
makes it recognize DEFINE_nnn and FLAGS_nnn conventions from the shflags
library.  We don't want to enable shflags on all files (not ebuilds, for
example), so plumb it through as an option instead of unconditionally
turning it on.

Bug: 898564
Test: make testrun; make test
Change-Id: I045d5abfe33fab9d686248a551732450256cf3ba
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1669194
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23821}