[Monorail] Rate limit based on request duration.
Generalize the current 1-count penalty for slow requests to be 1-count
per second of request time.  This enables us to raise the QPM rate limit
without risking that API users will DoS us with slow requests.

This change should be safe to deploy and roll back, even though this
version uses the same memcache entries as previous app versions.
That's because 95% of our API requests are under 2s, so they will
count the same, and most of the other 5% will count with a value
that is close enough to the old way of counting that the new higher
default rate limit will not be reached.

R=ehmaldonado@chromium.org

Bug: monorail:3262
Change-Id: Iccd3db082820377a929ad5d8e6f3ac1040bfef85
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1670048
Commit-Queue: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Edward Lesmes <ehmaldonado@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23887}