[tricium] Allow newline at end of commit message when checking skip footer flag
A footer flag like "Tricium: Skip" is supposed to make Tricium
skip analysis for a patchset, but this doesn't actually work now.

Upon closer inspection, I see that the footer flags are not
currently examined because of a bug in footer field extraction
that doesn't handle a commit message that ends in "\n".

Bug: 891090
Change-Id: Ibe484f0f54d539f065a125313f23c0c2b1d340fb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1672476
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23899}