[Monorail] Front-end support for federated blocking/blockedon references
This CL adds the ability to mark a bug as blocking or blocked on a
Google IssueTracker (b/) bug.

The plan for rolling this feature out is to add it to production
and ensure it works, but not to publicize it until the component
to update corresponding b/ bugs is written.

Bug: monorail:5639, monorail:5728
Change-Id: Id187de2d0b90fb1561dc5f346cc8d051ed08b5bb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1639678
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23927}