[tricium] Handle workflow runs with no functions
Currently, if a run is started that has no functions (e.g. because
there are no selected for a project, or they're all filtered out
by path filters), then the run is stuck in running state. This
CL would make it so that the run is immediately marked as done,
and a warning is logged.

Bug: 953006
Change-Id: I58d59389c33f91ee3ca12b17f72db36f6f937afb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1674329
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23952}