[Monorail] Spam: prevent flagged comments from counting toward threshold
There's a spam flag threshold number in settings (currently set to 5), above
that number of spam flags from non-admins, it marks any issue as spam. There
was a bug in that code that included comments in counting towards an issue's
spam threshold.

This is not the root cause of monorail:4027 or monorail:5781, however due
to the high number of false positives, this CL also effectively disables
the spam threshold flagging feature by setting the threshold to 1000.

I believe I discovered the root cause of old issues getting flagged as spam.
The SpamReport table does not enforce spam reports for a (user_id, comment_id,
issue_id) tuple to be unique, meaning that any user can replay their spam
report request and create new rows in that table. Combined with comments
originally counting toward that number (which this CL fixes), it ate up the
threshold quickly.

Bug: monorail:5813, monorail:5781
Change-Id: If3420eebd3ea15c8c0a4ef6565878e496b60020d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/1676465
Commit-Queue: Jeff Carpenter <jeffcarp@chromium.org>
Reviewed-by: Jason Robbins <jrobbins@chromium.org>
Cr-Commit-Position: refs/heads/master@{#23979}